backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
SQLAlchemy==2.0.23
python-multipart==0.0.6
pyyaml==6.0.1
synthetic-data-kit==0.0.3
pytube==15.0.0
youtube-transcript-api==0.6.1
pytz==2023.3
psutil==5.9.5
requests==2.31.0
